Adding undef. tests. (#36)

+++ src/streamvbyte_zigzag.c
@@ -1,8 +1,8 @@
 #include "streamvbyte_zigzag.h"
 
 static inline
-uint32_t _zigzag_encode_32 (int32_t val) {
-	return (val + val) ^ (val >> 31);
+uint32_t _zigzag_encode_32 (uint32_t val) {
+	return (val + val) ^ ((int32_t)val >> 31);
 }
 
 void zigzag_encode(const int32_t * in, uint32_t * out, size_t N) {
